used new drivers created for scheduler which have a dedicated function for each control signal instead of generic send control and dedicated function to transfer number of tasks+fix to number of tasks passed to the scheduler (prev was maxRTTasks, now number of tasks created in the system only, as it should be)

+++ portable.h
@@ -184,7 +184,7 @@ size_t xPortGetMinimumEverFreeHeapSize( void ) PRIVILEGED_FUNCTION;
  */
 void xPortScheduleNewTask(void) PRIVILEGED_FUNCTION;
 
-BaseType_t xPortInitScheduler( u16 numberOfTasks, void* pxRTTasksList, void* orderedDeadlineQTaskNums, void* orderedActivationQTaskNums, void* orderedDeadlineQPayload, void* orderedActivationQPayload, u32* pxCurrentTCBPtr) PRIVILEGED_FUNCTION;
+BaseType_t xPortInitScheduler( u32 numberOfTasks, void* pxRTTasksList, void* orderedDeadlineQTaskNums, void* orderedActivationQTaskNums, void* orderedDeadlineQPayload, void* orderedActivationQPayload, u32* pxCurrentTCBPtr) PRIVILEGED_FUNCTION;
 
 void newTaskHandler(void *HandlerRef) PRIVILEGED_FUNCTION;
